proptests: basic arithmetics eval invalid tests
